pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aquifer"
version = "0.1.0"
description = "Collect context, refreshing sources like RSS or YouTube etc."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"conduit",
	"feedparser>=6.0.12",
	"requests>=2.32.5",
	"requests-cache>=1.2.1",
	"diskcache>=5.6.3",
	"feedparser>=6.0.12",
	"google-genai>=1.56.0",
	"requests>=2.32.5",
]

# ── uv sources (local/editable deps) ───────────────────────────────────────────
[tool.uv.sources]
dbclients = { path = "../dbclients-project", editable = true }
conduit = { path = "../conduit-project", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/aquifer"]
sources = ["src"]
include = ["src/aquifer/py.typed"]

[tool.hatch.build.targets.sdist]
include = ["src/aquifer/py.typed"]
